+++ packages/behandling-aktivitetspenger/src/components/AktivitetspengerProsess.tsx
@@ -15,7 +15,7 @@ import { BeregningProsessStegInitPanel } from './prosess/BeregningProsessStegIni
 import { ForutgåendeMedlemskapInitPanel } from './prosess/ForutgåendeMedlemskapInitPanel';
 import { InngangsvilkårInitPanel } from './prosess/InngangsvilkårInitPanel';
 import { VedtakProsessStegInitPanel } from './prosess/VedtakProsessStegInitPanel';
-import { useProsessmotor } from './Prossesmotor';
+import { useProsessmotor } from './Prosessmotor';
 
 interface OwnProps {
   api: AktivitetspengerApi;
